requirements.txt
streamlit
openai
python-dotenv
faiss-cpu
langchain
pymupdf
python-docx
pdfplumber
langchain-community
tiktoken